I'm good at breaking shit. should be fixed now
forgot to sync first this morning, as I had made a small change on
Prose.io last night from my phone. DUMB! anyways, broke the HEAD of this
branch, so I deleted the local copy and recloned. The change from last
night is included, which was re-inlining the project link in Koffee
House.