logo/nav tweaks
The changes I made here accomplish a couple things: it prevents random
"height-hopping" when hovering the logo after scrolling around the page
a bit (the nav bar grew and shrunk by a pixel or so randomly), it's
easier to control the height and <main> padding, smoother animation, and
it also allowed me to refactor the code a bit better. Also vertically
centered everything better, too.