quick fix for candy columns
Since .candy-columns-- is intended to be used only in an article
context, I don't see what's wrong with removing the bottom margin so
that I can use it within an article. We'll probably need to remove
margin-bottom from any <div> within an article, though. Also, trying to
fix error pages.