weird, see commit desc
so, apparently NOT having a parallax header <div> causes a misfire in
index.min.js. For the time being, I've moved the parallax code to the
END of this file, so that it will still error out, but not before the
rest of index.min.js is executed. Previously, the error caused ALL
scripts to be killed. This issue still needs to be fixed!!!